refactor(web): extract catalog import flow
